`processor` - Use `AyonShotgridHub` in `processor.handlers` Removed outdated handlers and confomed all the logic in two, which make use of the `AyonShotgridHub` simplifying the overall logic.
